pyproject.toml
[project]
name = "tardis-data-downloader"
version = "0.1.0"
description = "Tardis data download and loading"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Da-Wei Lee"},
]
keywords = ["tardis", "crypto", "data"]
dependencies = [
    "fire>=0.7.1",
    "loguru>=0.7.3",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "tardis-dev>=2.1.2",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipdb>=0.13.13",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "nbconvert>=7.16.6",
]

[project.urls]
Repository = "https://github.com/daviddwlee84/Tardis-Downloader"

# By default hatchling will build the package in the src/package_name_with_underscores directory
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# https://github.com/astral-sh/uv/issues/6925
[tool.uv]
index-url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"


[project.scripts]
tardis-data-downloader = "tardis_data_downloader.cli.tardis_data_downloader:main"
td-fire = "tardis_data_downloader.cli.fire_download_wrapper:main"
td-ui = "tardis_data_downloader.ui.streamlit_runner:main"
# td-streamlit = "tardis_data_downloader.ui.streamlit_direct:main"
